Move async stream utilities to common folder as requested
